+++ NEWS.md
@@ -1,7 +1,8 @@
 # plotly (development version)
 
+## Improvements
 
-
+* Create the proper `bundleTraceMap` list as internal package data object directly from the upstream JS file when plotly.js is updated. As a consequence, it's now possible to use the [`strict` bundle](https://github.com/plotly/plotly.js/blob/master/dist/README.md#plotlyjs-strict) via `plotly::partial_bundle(type = "strict")`.
 
 # plotly 4.12.0